fix the mass calculation to exit gracefully.
Update also changes the mass calculation significantly - now more flexibile and easy to add new categories to, and does not require an entire category to be complete. vNotifies use of individual DTOs missing exchanges. Added ability to reload the entire UI and retain tab position so it doesn't look jank.