improve logging
this commit cleans up some sneaky prints, as well as converts all of the logging.infos in canvas to logging.debug. it also adds kwargs to main so you can set the run parameters to --level=INFO to ignore the debug logs entirely, because they're generally just clutter if you're not actively debugging.